refactor: move messageList to it's own component
